refactored backend directory
