+++ backend/requirements.txt
@@ -5,7 +5,6 @@ email-validator>=2.0.0
 fastapi>=0.115.8
 httpx>=0.28.1
 passlib>=1.7.4
-psycopg2>=2.2.0
 psycopg2-binary>=2.9.10
 pydantic>=2.10.6
 pydantic_core>=2.27.2
