+++ backend/v1/app/routes/system/__init__.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 
-from backend.v1.app.api.system.health_check import router as health_router
+from backend.v1.app.routes.system.health_check import router as health_router
 
 
 router = APIRouter()
