+++ backend/v1/app/routes/users/__init__.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 
-from backend.v1.app.api.users.users import router as user_router
-from backend.v1.app.api.users.stylist import router as stylist_router
+from backend.v1.app.routes.users.users import router as user_router
+from backend.v1.app.routes.users.stylist import router as stylist_router
 
 
 router = APIRouter()
