+++ backend/v1/app/server/__init__.py
@@ -1,0 +1,1 @@
+"""CurlUp Backend Server Module."""
